adicionar verificação de permissão para atualização e exclusão de usuários com papel ADMIN

+++ src/modules/users/users.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -7,6 +8,7 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersAbstractRepository } from './repositories/users-abstract.repository';
 import { UserEntity } from './entities/user.entity';
+import { JwtPayload } from 'src/jwt/jwt-token.service';
 
 @Injectable()
 export class UsersService {
@@ -47,10 +49,19 @@ export class UsersService {
     };
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(
+    id: number,
+    updateUserDto: UpdateUserDto,
+    currentUser: JwtPayload,
+  ) {
     const targetUser = await this.usersRepository.findUserById(id);
     if (!targetUser) {
       throw new NotFoundException('User not found');
+    }
+
+    const isPermitted = this.adminCheck(currentUser, targetUser);
+    if (!isPermitted) {
+      throw new ForbiddenException('You are not allowed to update this user');
     }
 
     Object.assign(targetUser, updateUserDto);
@@ -65,15 +76,31 @@ export class UsersService {
     };
   }
 
-  async remove(id: number) {
+  async remove(id: number, currentUser: JwtPayload) {
     const targetUser = await this.usersRepository.findUserById(id);
     if (!targetUser) {
       throw new NotFoundException('User not found');
+    }
+
+    const isPermitted = this.adminCheck(currentUser, targetUser);
+    if (!isPermitted) {
+      throw new ForbiddenException('You are not allowed to delete this user');
     }
 
     const { password, ...user } = await this.usersRepository.deleteUser(id);
     return {
       user,
     };
+  }
+
+  private adminCheck(currentUser: JwtPayload, targetUser: UserEntity) {
+    if (
+      targetUser.role === 'ADMIN' &&
+      currentUser.role !== 'SUPER_ADMIN' &&
+      targetUser.id !== currentUser.sub
+    ) {
+      return false;
+    }
+    return true;
   }
 }
